refactor: HandlerAdapters 클래스로 책임 분리
